feat: 아티스트 대시보드 뷰에서의 이 달의 track pagination 반환 API 추가

+++ src/main/java/com/github/bluekey/dto/response/artist/ArtistMonthlyTrackListResponseDto.java
@@ -13,13 +13,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ArtistMonthlyTrackListResponseDto {
     @Schema(description = "총 아이템 개수", example = "300")
-    private Long totalItems;
+    private int totalItems;
 
     @Schema(description = "앨범 정보")
     private List<ArtistMonthlyTrackListDto> contents;
 
     @Builder
-    public ArtistMonthlyTrackListResponseDto(final Long totalItems, final List<ArtistMonthlyTrackListDto> contents) {
+    public ArtistMonthlyTrackListResponseDto(final int totalItems, final List<ArtistMonthlyTrackListDto> contents) {
         this.totalItems = totalItems;
         this.contents = contents;
     }
